UpdateQauntityAndPrices

+++ Models/Assay.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; } 
         public string Name { get; set; }
         public string Description { get; set; }
-        public int MinimumPrice {  get; set; }
-        public int MaximumPrice { get; set; }
+        public float MinimumPrice {  get; set; }
+        public float MaximumPrice { get; set; }
         public string Unit {  get; set; }
 
     }
@@ -14,8 +14,8 @@
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public int MinimumPrice { get; set; }
-        public int MaximumPrice { get; set; }
+        public float MinimumPrice { get; set; }
+        public float MaximumPrice { get; set; }
         public string Unit { get; set; }
     }
 }
